backend/requirements.txt
torch
Pillow
torchvision
fastapi
uvicorn
